requirements.txt
black==22.6.0
certifi==2022.6.15
click==8.1.3
elastic-transport==8.1.2
elasticsearch==8.3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dotenv==0.20.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
